# CodigoFuenteAplicacion/src/main/resources/META-INF/openapi.yaml
openapi: 3.0.0
info:
  title: Customers API
  description: API para la gestión de clientes - Grupo 4 Automatizacion de Procesos
  version: "1.0.0"
servers:
  - url: http://localhost:8082
    description: Grupo 4 Automatizacion de Procesos
paths:
  /customers/list-all:
    post:
      summary: Listar todos los clientes
      description: Obtiene una lista de todos los clientes registrados.
      responses:
        '200':
          description: Listado OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    nombres:
                      type: string
                      example: Juan
                    apellidos:
                      type: string
                      example: Pérez
                    cedula:
                      type: string
                      example: 0102030405
                    correo:
                      type: string
                      format: email
                      example: juan.perez@example.com
                    telefono:
                      type: string
                      example: 0987654321
                    fechaNacimiento:
                      type: string
                      format: date-time
                      example: 1990-01-15T05:00:00.000+00:00
  /customers/save:
    post:
      summary: Guardar un nuevo cliente
      description: Guarda un nuevo cliente con los datos proporcionados.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombres:
                  type: string
                  example: Juan
                apellidos:
                  type: string
                  example: Perez
                cedula:
                  type: string
                  example: 0106487846
                correo:
                  type: string
                  format: email
                  example: lcabrera@est.ups.edu.ec
                telefono:
                  type: string
                  example: 0991234567
                fechaNacimiento:
                  type: string
                  format: date
                  example: 1990-05-15
      responses:
        '200':
          description: Guardado OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 6
                  nombres:
                    type: string
                    example: Juan
                  apellidos:
                    type: string
                    example: Perez
                  cedula:
                    type: string
                    example: 0106487846
                  correo:
                    type: string
                    format: email
                    example: lcabrera@est.ups.edu.ec
                  telefono:
                    type: string
                    example: 0991234567
                  fechaNacimiento:
                    type: string
                    format: date-time
                    example: 1990-05-15T00:00:00.000+00:00
        '400':
          description: Error en la solicitud aplica para correo y cedula
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: La cedula/correo del cliente es invalida

  /customers/update:
    put:
      summary: Actualizar un cliente existente
      description: Actualiza los datos de un cliente existente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                nombres:
                  type: string
                  example: Romel
                apellidos:
                  type: string
                  example: Avila
                cedula:
                  type: string
                  example: 0102030405
                correo:
                  type: string
                  format: email
                  example: ravila@est.ups.edu.ec
                telefono:
                  type: string
                  example: 1234567848
                fechaNacimiento:
                  type: string
                  format: date-time
                  example: 2001-12-20T05:00:00.000+00:00
      responses:
        '200':
          description: Cliente actualizado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  nombres:
                    type: string
                    example: Romel
                  apellidos:
                    type: string
                    example: Avila
                  cedula:
                    type: string
                    example: 0102030405
                  correo:
                    type: string
                    format: email
                    example: ravila@est.ups.edu.ec
                  telefono:
                    type: string
                    example: 1234567848
                  fechaNacimiento:
                    type: string
                    format: date-time
                    example: 2001-12-20T05:00:00.000+00:00
  /customers/delete/{id}:
    delete:
      summary: Eliminar un cliente
      description: Elimina un cliente por su ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del cliente a eliminar.
      responses:
        '200':
          description: Cliente eliminado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Customer successfully deleted